+++ src/main.ts
@@ -33,6 +33,7 @@ import {
     setupTokenHUDResetMovementBtn,
 } from "./token/setup";
 import { setupTokenHUDAppearanceControls } from "./token/hudAppearance";
+import { setupKeyboardMovementAnimation } from "./token/keyboardMovement";
 import { isLegalStep } from "./token/trail/isLegalStep";
 import { getRowsCol } from "./utils/rowsCols";
 import { getOccupiedTiles } from "./utils/tiles";
@@ -51,6 +52,7 @@ Hooks.once("init", () => {
     setupSupressCanvasPanWhileDragging();
     setupTokenHUDResetMovementBtn();
     setupTokenHUDAppearanceControls();
+    setupKeyboardMovementAnimation();
     setupMovementHistory();
     bakeFont();
 });